Remove obsolete WPF Panel2D rendering factory
