+++ WindowsNetProjects/OasisEditor/OasisEditor.Tests/Panel2DRoundTripTests.cs
@@ -1,5 +1,4 @@
 using Xunit;
-using System.Windows;
 using System.Collections.ObjectModel;
 
 namespace OasisEditor.Tests;
@@ -1504,29 +1503,6 @@ public sealed class Panel2DRoundTripTests
         Assert.Equal("Second Renamed", secondDocument.GetPanelElements().Single().Name);
     }
 
-    [Theory]
-    [InlineData(0.1, 9.9, 0.0, 10.0)]
-    [InlineData(14.9, 15.1, 10.0, 20.0)]
-    [InlineData(25.0, 35.0, 20.0, 40.0)]
-    [InlineData(-4.9, -5.1, 0.0, -10.0)]
-    public void SnapPointToGrid_RoundsToNearestTenPixels(double x, double y, double expectedX, double expectedY)
-    {
-        var snapped = PanelToolPlacementController.SnapPointToGrid(new Point(x, y));
-
-        Assert.Equal(expectedX, snapped.X);
-        Assert.Equal(expectedY, snapped.Y);
-    }
-
-    [Fact]
-    public void SnapPointToGrid_PreservesExactGridCoordinates()
-    {
-        var point = new Point(40.0, -30.0);
-
-        var snapped = PanelToolPlacementController.SnapPointToGrid(point);
-
-        Assert.Equal(40.0, snapped.X);
-        Assert.Equal(-30.0, snapped.Y);
-    }
 
     private static DocumentTabViewModel CreatePanelDocument(params PanelElementModel[] elements)
     {
