+++ WindowsNetProjects/OasisEditor/OasisEditor/Rendering/LampElementRenderer.cs
@@ -357,29 +357,7 @@ internal sealed class LampElementRenderer : IPanelElementRenderer
 
     private static bool TryResolveAssetPath(string? assetPath, out string resolvedPath)
     {
-        resolvedPath = string.Empty;
-        if (string.IsNullOrWhiteSpace(assetPath))
-        {
-            return false;
-        }
-
-        var candidate = assetPath.Trim();
-        if (Path.IsPathRooted(candidate))
-        {
-            resolvedPath = candidate;
-            return true;
-        }
-
-        if (string.IsNullOrWhiteSpace(PanelElementFactory.ProjectDirectoryPath))
-        {
-            return false;
-        }
-
-        var relativePath = candidate
-            .Replace('/', Path.DirectorySeparatorChar)
-            .Replace('\\', Path.DirectorySeparatorChar);
-        resolvedPath = Path.GetFullPath(Path.Combine(PanelElementFactory.ProjectDirectoryPath, relativePath));
-        return true;
+        return ProjectAssetPathResolver.TryResolveAssetPath(assetPath, out resolvedPath);
     }
 
     private static SKColor Lerp(SKColor from, SKColor to, double t)
