+++ WindowsNetProjects/OasisEditor/OasisEditor/Views/SkiaFaceEditView.xaml.cs
@@ -385,29 +385,7 @@ public partial class SkiaFaceEditView : UserControl
 
     private static bool TryResolveAssetPath(string? assetPath, out string resolvedPath)
     {
-        resolvedPath = string.Empty;
-        if (string.IsNullOrWhiteSpace(assetPath))
-        {
-            return false;
-        }
-
-        var candidate = assetPath.Trim();
-        if (Path.IsPathRooted(candidate))
-        {
-            resolvedPath = candidate;
-            return true;
-        }
-
-        if (string.IsNullOrWhiteSpace(PanelElementFactory.ProjectDirectoryPath))
-        {
-            return false;
-        }
-
-        var relativePath = candidate
-            .Replace('/', Path.DirectorySeparatorChar)
-            .Replace('\\', Path.DirectorySeparatorChar);
-        resolvedPath = Path.GetFullPath(Path.Combine(PanelElementFactory.ProjectDirectoryPath, relativePath));
-        return true;
+        return ProjectAssetPathResolver.TryResolveAssetPath(assetPath, out resolvedPath);
     }
 
     private void OnFaceSkiaSurfaceMouseDown(object sender, MouseButtonEventArgs eventArgs)
